+++ src/accessories/index.ts
@@ -2,3 +2,4 @@ export * from './BaseAccessory';
 export * from './DimmableLightAccessory';
 export * from './LightAccessory';
 export * from './SwitchAccessory';
+export * from './BlindsAccessory';